+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_help.py
@@ -540,3 +540,30 @@ helps['storage table policy'] = """
     type: group
     short-summary: Manage shared access policies of a storage table.
 """
+
+helps['storage account network-rule'] = """
+    type: group
+    short-summary: Manage network rules.
+"""
+
+helps['storage account network-rule add'] = """
+    type: command
+    short-summary: Add a network rule.
+    long-summary: >
+        Rules can be created for an IPv4 address, address range (CIDR format), or a virtual network subnet.
+    examples:
+        - name: Create a rule to allow a specific address-range.
+          text: az storage account network-rule add -g myRg --account-name mystorageaccount --ip-address 23.45.1.0/24
+        - name: Create a rule to allow access for a subnet.
+          text: az storage account network-rule add -g myRg --account-name mystorageaccount --vnet myvnet --subnet mysubnet
+"""
+
+helps['storage account network-rule list'] = """
+    type: command
+    short-summary: List network rules.
+"""
+
+helps['storage account network-rule remove'] = """
+    type: command
+    short-summary: Remove a network rule.
+"""