+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_params.py
@@ -35,7 +35,7 @@ from ._validators import \
      process_file_download_namespace,
      process_metric_update_namespace, process_blob_copy_batch_namespace,
      get_source_file_or_blob_service_client, process_blob_source_uri,
-     get_char_options_validator)
+     get_char_options_validator, validate_bypass, validate_subnet)
 
 
 DeleteSnapshot, BlockBlobService, \
@@ -359,6 +359,15 @@ with CommandContext('storage account update') as c:
             g.reg_extra_arg('encryption_key_vault', help='The Uri of the KeyVault')
             g.reg_extra_arg('encryption_key_version', help='The version of the KeyVault key')
 
+with VersionConstraint(ResourceType.MGMT_STORAGE, min_api='2017-06-01') as c:
+    for item in ['create', 'update']:
+        register_cli_argument('storage account {}'.format(item), 'bypass', nargs='+', validator=validate_bypass, arg_group='Network Rule', help='Bypass traffic for space-separated uses.', **model_choice_list(ResourceType.MGMT_STORAGE, 'Bypass'))
+        register_cli_argument('storage account {}'.format(item), 'default_action', arg_group='Network Rule', help='Default action to apply when no rule matches.', **model_choice_list(ResourceType.MGMT_STORAGE, 'DefaultAction'))
+
+    register_cli_argument('storage account network-rule', 'storage_account_name', account_name_type)
+    register_cli_argument('storage account network-rule', 'ip_address', help='IPv4 address or CIDR range.')
+    register_cli_argument('storage account network-rule', 'subnet', help='Name or ID of subnet. If name is supplied, `--vnet-name` must be supplied.')
+    register_cli_argument('storage account network-rule', 'vnet_name', help='Name of a virtual network.', validator=validate_subnet)
 
 register_cli_argument('storage account keys renew', 'key_name', options_list=('--key',), help='The key to regenerate.', validator=validate_key, **enum_choice_list(list(storage_account_key_options.keys())))
 register_cli_argument('storage account keys renew', 'account_name', account_name_type, id_part=None)